+++ HspppTest/ApiCompileTest.cpp
@@ -666,9 +666,9 @@ namespace compile_test {
         [[maybe_unused]] std::string s4 = hsppp::str(0.0);
 
         // strlen
-        [[maybe_unused]] int len1 = hsppp::strlen("Hello");
-        [[maybe_unused]] int len2 = hsppp::strlen("");
-        [[maybe_unused]] int len3 = hsppp::strlen(std::string("日本語"));  // マルチバイト
+        [[maybe_unused]] int64_t len1 = hsppp::strlen("Hello");
+        [[maybe_unused]] int64_t len2 = hsppp::strlen("");
+        [[maybe_unused]] int64_t len3 = hsppp::strlen(std::string("日本語"));  // マルチバイト
     }
 
     // ============================================================
@@ -703,14 +703,14 @@ namespace compile_test {
     // ============================================================
     void test_string_functions() {
         // instr - 文字列の検索
-        [[maybe_unused]] int pos1 = hsppp::instr("ABCDEF", "CD");          // 2
-        [[maybe_unused]] int pos2 = hsppp::instr("ABCDEF", 2, "CD");       // 0 (2を起点)
-        [[maybe_unused]] int pos3 = hsppp::instr("ABCDEF", "XY");          // -1
-        [[maybe_unused]] int pos4 = hsppp::instr("ABCDEF", -1, "CD");      // -1 (負のインデックス)
-        [[maybe_unused]] int pos5 = hsppp::instr("", "ABC");               // -1
-        [[maybe_unused]] int pos6 = hsppp::instr("ABCABC", 3, "ABC");      // 0 (3を起点)
-        [[maybe_unused]] int pos7 = hsppp::instr("Hello World", 0, "World"); // 6
-        [[maybe_unused]] int pos8 = hsppp::instr("ABC", "");               // 0 (空文字検索)
+        [[maybe_unused]] int64_t pos1 = hsppp::instr("ABCDEF", "CD");          // 2
+        [[maybe_unused]] int64_t pos2 = hsppp::instr("ABCDEF", 2, "CD");       // 0 (2を起点)
+        [[maybe_unused]] int64_t pos3 = hsppp::instr("ABCDEF", "XY");          // -1
+        [[maybe_unused]] int64_t pos4 = hsppp::instr("ABCDEF", -1, "CD");      // -1 (負のインデックス)
+        [[maybe_unused]] int64_t pos5 = hsppp::instr("", "ABC");               // -1
+        [[maybe_unused]] int64_t pos6 = hsppp::instr("ABCABC", 3, "ABC");      // 0 (3を起点)
+        [[maybe_unused]] int64_t pos7 = hsppp::instr("Hello World", 0, "World"); // 6
+        [[maybe_unused]] int64_t pos8 = hsppp::instr("ABC", "");               // 0 (空文字検索)
 
         // strmid - 文字列の一部を取り出す
         [[maybe_unused]] std::string mid1 = hsppp::strmid("ABCDEF", 1, 3);   // "BCD"
@@ -761,14 +761,14 @@ namespace compile_test {
 
         // strrep - 文字列の置換
         std::string repStr = "aaa bbb aaa ccc";
-        [[maybe_unused]] int repCount1 = hsppp::strrep(repStr, "aaa", "XXX");    // 2 (置換回数)
+        [[maybe_unused]] int64_t repCount1 = hsppp::strrep(repStr, "aaa", "XXX");    // 2 (置換回数)
         // repStr は "XXX bbb XXX ccc" になっている
         std::string repStr2 = "ABCABC";
-        [[maybe_unused]] int repCount2 = hsppp::strrep(repStr2, "ABC", "X");     // 2
+        [[maybe_unused]] int64_t repCount2 = hsppp::strrep(repStr2, "ABC", "X");     // 2
         std::string repStr3 = "Hello";
-        [[maybe_unused]] int repCount3 = hsppp::strrep(repStr3, "XYZ", "");      // 0 (見つからない)
+        [[maybe_unused]] int64_t repCount3 = hsppp::strrep(repStr3, "XYZ", "");      // 0 (見つからない)
         std::string repStr4 = "";
-        [[maybe_unused]] int repCount4 = hsppp::strrep(repStr4, "A", "B");       // 0 (空文字列)
+        [[maybe_unused]] int64_t repCount4 = hsppp::strrep(repStr4, "A", "B");       // 0 (空文字列)
 
         // getstr - バッファから文字列読み出し
         std::string strBuf = "ABC,DEF,GHI";
@@ -1054,20 +1054,20 @@ namespace compile_test {
         std::vector<uint8_t> vecBuf(64, 0);
         
         // bload シグネチャ
-        [[maybe_unused]] int bl1 = bload("test.bin", strBuf);
-        [[maybe_unused]] int bl2 = bload("test.bin", strBuf, 32);
-        [[maybe_unused]] int bl3 = bload("test.bin", strBuf, 32, 0);
-        [[maybe_unused]] int bl4 = bload("test.bin", vecBuf);
-        [[maybe_unused]] int bl5 = bload("test.bin", vecBuf, 32);
-        [[maybe_unused]] int bl6 = bload("test.bin", vecBuf, 32, 0);
+        [[maybe_unused]] int64_t bl1 = bload("test.bin", strBuf);
+        [[maybe_unused]] int64_t bl2 = bload("test.bin", strBuf, 32);
+        [[maybe_unused]] int64_t bl3 = bload("test.bin", strBuf, 32, 0);
+        [[maybe_unused]] int64_t bl4 = bload("test.bin", vecBuf);
+        [[maybe_unused]] int64_t bl5 = bload("test.bin", vecBuf, 32);
+        [[maybe_unused]] int64_t bl6 = bload("test.bin", vecBuf, 32, 0);
 
         // bsave シグネチャ
-        [[maybe_unused]] int bs1 = bsave("test.bin", strBuf);
-        [[maybe_unused]] int bs2 = bsave("test.bin", strBuf, 32);
-        [[maybe_unused]] int bs3 = bsave("test.bin", strBuf, 32, 0);
-        [[maybe_unused]] int bs4 = bsave("test.bin", vecBuf);
-        [[maybe_unused]] int bs5 = bsave("test.bin", vecBuf, 32);
-        [[maybe_unused]] int bs6 = bsave("test.bin", vecBuf, 32, 0);
+        [[maybe_unused]] int64_t bs1 = bsave("test.bin", strBuf);
+        [[maybe_unused]] int64_t bs2 = bsave("test.bin", strBuf, 32);
+        [[maybe_unused]] int64_t bs3 = bsave("test.bin", strBuf, 32, 0);
+        [[maybe_unused]] int64_t bs4 = bsave("test.bin", vecBuf);
+        [[maybe_unused]] int64_t bs5 = bsave("test.bin", vecBuf, 32);
+        [[maybe_unused]] int64_t bs6 = bsave("test.bin", vecBuf, 32, 0);
 
         // dialog（シグネチャのみ確認）
         dialog("メッセージ");
